* Symtab.java: More code to handle abstract and native methods.
* java.g: Should handle abstract and native methods now.

* AssertTools.java: Added method to translate a package name into
a directory name and create the directories relative to a
destination directory.  This will allow me to instrument files
into another directory rather than just changing the extension and
confusing compilers.

* Main.java: Added code to check modification times of files
before parsing, so we don't parse extra files.

* AssertMethod.java: Added flag to tell if this method is abstract/native.

* CodeGenerator.java: private methods have private pre and post
condition check methods.  Also the check for a super method isn't
done if the method is private.


git-svn-id: https://jon/svn/JonsAssert/trunk@36 94f1d44e-47f0-0310-857f-9ad392bb85a4